requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
SQLAlchemy
psycopg2==2.9.3
python-dotenv==0.21.0
SQLAlchemy==1.4.41
pytest
pytest-cov
